[Refactor] 롤 리그 조회 및 즐겨찾기 로직 변경 등 - 롤 팀 즐겨찾기로 인한 정렬 로직을 분리하여 불필요한 리그 fetch 방지 - 컴포넌트 폴더 구조 세분화 및 파일 경로 alias 적용  > 공통: /common  > 롤 관련: /lol  > 인증: /auth
[Fix] 롤 리그 셀렉터 드롭다운이 위쪽으로 열리도록 스타일 수정
