+++ src/components/lol/FavoriteTeamButton.js
