+++ src/components/lol/LeagueAndTeamSelector.js
@@ -1,17 +1,19 @@
 import {useEffect, useState} from "react";
-import FavoriteTeamButton from "@components/FavoriteTeamButton";
+import FavoriteTeamButton from "@components/lol/FavoriteTeamButton";
 import {useCalandar} from "@/context/CalandarContext";
 import {useAuth} from "@/context/AuthContext.js";
-import LeagueDropdown from "@components/LeagueDropdown.js";
+import LeagueDropdown from "@components/lol/LeagueDropdown.js";
 
-const TeamSelector = () => {
+const LeagueAndTeamSelector = () => {
     const [leagues, setLeagues] = useState([]);
     const [selectedLeagueId, setSelectedLeagueId] = useState('');
-    const [teams, setTeams] = useState([]);
+    const [rawTeams, setRawTeams] = useState([]); // 👈 fetch 결과만 보관
+    const [teams, setTeams] = useState([]);       // 👈 정렬된 최종 데이터
+
     const {favoriteTeamIds} = useCalandar();
     const {userId} = useAuth();
 
-
+    // 🎯 리그 목록 가져오기
     useEffect(() => {
         const fetchLeagues = async () => {
             try {
@@ -19,7 +21,6 @@ const TeamSelector = () => {
                 const data = await res.json();
                 setLeagues(data);
 
-                // ✅ 여기서 초기 선택값 설정
                 if (!selectedLeagueId && data.length > 0) {
                     setSelectedLeagueId(data[0].id);
                 }
@@ -31,37 +32,41 @@ const TeamSelector = () => {
         fetchLeagues();
     }, []);
 
+    // 🎯 리그 변경 시 팀 fetch
     useEffect(() => {
         const fetchTeams = async () => {
             try {
                 const res = await fetch(`/api/lol/teams?leagueId=${selectedLeagueId}`);
                 const data = await res.json();
-
-                // 🎯 즐겨찾기 팀 먼저 정렬
-                const sortedData = [...data].sort((a, b) => {
-                    const aIndex = favoriteTeamIds.indexOf(a.teamId);
-                    const bIndex = favoriteTeamIds.indexOf(b.teamId);
-
-                    // 둘 다 즐겨찾기에 있음 → 순서 비교
-                    if (aIndex !== -1 && bIndex !== -1) return aIndex - bIndex;
-
-                    // 하나만 즐겨찾기임 → 즐겨찾기 먼저
-                    if (aIndex !== -1) return -1;
-                    if (bIndex !== -1) return 1;
-
-                    return 0; // 둘 다 즐겨찾기 아님 → 원래 순서 유지
-                });
-
-                setTeams(sortedData);
+                setRawTeams(data); // fetch만 담당
             } catch (e) {
                 console.error("팀 정보를 불러오는 데 실패했습니다.", e);
             }
         };
 
-        fetchTeams();
-    }, [favoriteTeamIds, selectedLeagueId]);
+        if (selectedLeagueId) {
+            fetchTeams();
+        }
+    }, [selectedLeagueId]);
 
+    // 🎯 즐겨찾기 or rawTeams 변경 시 정렬
+    useEffect(() => {
+        if (!rawTeams.length) return;
 
+        const sortedData = [...rawTeams].sort((a, b) => {
+            const aIndex = favoriteTeamIds.indexOf(a.teamId);
+            const bIndex = favoriteTeamIds.indexOf(b.teamId);
+
+            if (aIndex !== -1 && bIndex !== -1) return aIndex - bIndex;
+            if (aIndex !== -1) return -1;
+            if (bIndex !== -1) return 1;
+            return 0;
+        });
+
+        setTeams(sortedData);
+    }, [favoriteTeamIds, rawTeams]);
+
+    // ✅ 즐겨찾기 팀 분리
     const favoriteTeams = teams.filter(team =>
         favoriteTeamIds.includes(team.teamId)
     );
@@ -71,29 +76,24 @@ const TeamSelector = () => {
 
     return (
         <div className="team-selector-wrapper">
-            {/* 🔽 리그 셀렉트박스 영역 */}
             <LeagueDropdown
                 leagues={leagues}
                 selectedLeague={selectedLeagueId}
                 onChange={setSelectedLeagueId}
             />
 
-            {/* ⭐ 즐겨찾기 고정 팀 영역 (리그 변경 시에도 유지) */}
-            {userId &&
-                (
-                    <div
-                        className={`favorite-teams-section scroll-hidden${favoriteTeams.length === 0 ? ' empty' : ''}`}
-                    >
-                        <div className="team-btn-container">
-                            {favoriteTeams.map(team => (
-                                <FavoriteTeamButton key={team.teamId} {...team} />
-                            ))}
-                        </div>
+            {userId && (
+                <div
+                    className={`favorite-teams-section scroll-hidden${favoriteTeams.length === 0 ? ' empty' : ''}`}
+                >
+                    <div className="team-btn-container">
+                        {favoriteTeams.map(team => (
+                            <FavoriteTeamButton key={team.teamId} {...team} />
+                        ))}
                     </div>
-                )
-            }
+                </div>
+            )}
 
-            {/* 📅 리그별 팀 목록 영역 */}
             <div
                 className={`nonfavorite-teams-section scroll-hidden${nonFavoriteTeams.length === 0 ? ' empty' : ''}`}
             >
@@ -107,4 +107,4 @@ const TeamSelector = () => {
     );
 };
 
-export default TeamSelector;
+export default LeagueAndTeamSelector;
